Use the chat role manager to get chat roles
